refactor: move birthdate component to web common
